move SCC tests to e2e
